backend/requirements.txt
flask
flask-cors
scikit-learn
numpy
pandas
nltk
python-dotenv
requestsi 